Make Azure Connection abstract
Have concrete objects for differing Azure connection types (Graph
API/Data Lake/etc.)
